requirements.txt
requests
websockets
wazirx_sapi_client